+++ packages/sources/tradingeconomics/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export * as markets from './markets'